NN code beginning to be implemented
Have temporarily solved an issue with the datasets not being able to be downloaded using the solution from canvas. Should revisit this to enusre we can download the files from locations that function but moving ahead for now. Additionally we are not using any convalutional layers yet but these may be added once the functionality is testable